Shifted running PillowFight command to another class
